Adjust xio_mt_client.c to match Ceph thread contexts.
To be completely consistent, the mt1 example needed to
connect after spawning the portal_worker_thread.

The symbol UNBOUND_OUTBOUND_THREAD is not defined.  Another way that the
Ceph client differs from this one is in starting a "client" (unbound)
context before creating the bound ones.  The conditional sections test
that case--but are disabled, because we reproduce the bad behavior without
spawning the extra context.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>