Adjust xio_mt_server.c to match Ceph thread contexts.
THIS change, unlike the prior client change, causes the hello
world client not to respond, apparently similarly to the behavior
I am seeing with the Ceph XioMessenger.

The only difference I -believe- to exist relative to the supplied
hello_world_mt1 is that the default event loop is bound in a spawned
thread context, rather than in main().

This could cause the default xio_bind to race with the portal context
xio_binds, and that actually may be a different change from the one
that actually happens in the Ceph Messenger--but since the user-visible
result is similar, it seems relevant.

In my first version of this change, there was a race to start default and
portal threads.  Now, main() waits for the acceptor_thread to xio_bind.

The symbol UNBOUND_OUTBOUND_THREAD is not defined.  Another way that the
Ceph client differs from this one is in starting a "client" (unbound)
context before creating the bound ones.  The conditional sections test
that case--but are disabled, because we reproduce the bad behavior without
spawning the extra context.

Signed-off-by: Matt Benjamin <matt@linuxbox.com>