Add hello_world1.
This example attempts to verify that Accelio permits
applications to relibly free resources it has allocated and
presented to libxio.

It appears that xio_client.c l. 119 and l. 134 show incorrect
behavior (change the preprocessor directives to see this).

The key expectation being violated is that xio response messages
being returned to callers in the on_msg callback are not
intact, and may not be freed (but xio MUST allow the application
to free these resources, as xio did not allocate them).

Signed-off-by: Matt Benjamin <matt@linuxbox.com>